Reintroduce unsafe DMA transfer methods, remove ptr checks
Besides the lifetime management, which is described in the
set_[source|destination] transfer methods, there are other reasons
why creating a Transfer is unsafe. This commit makes the methods
unsafe again.

Also, remove the pointer alignment check in buffer_circular. If the
start pointer points at an offset, it's not aligned to the byte
boundary. The safety documentation describes the alignment
requirements to use the circular buffer transfers.