Remove hard-coded DMA, DMAMUX addresses; on_error
Similar to the USB device driver, the DMA and DMAMUX addresses change
once we consider the 11xx MCUs. This commit refactors the driver so that
it can support arbitrary DMA and DMAMUX locations.

Construct a `Dma` with pointers to the peripherals. Then, use the `Dma`
object to produce channels. It's very similar to what we had before;
only now, the pointers are configured by the user. I opted for a const
function, so that a HAL / higher-level driver can allocate the DMA
driver in a constant. This lets us move the maximum number of channels
into the constant.

This requires that we attach the on_interrupt function to
`Dma`. It behaves as before. This commit removes on_error, rather than
refactoring it. It's not clear how to use this function, since (1) we're
not supporting an API to enable error interrupts per channel, and (2)
the Transfer future isn't really designed for a 1:N error:channel
interrupt. I added a comment in the Transfer future to describe how it
expects errors to work.